% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bps_from_s3.R
\name{read_bps_from_s3}
\alias{read_bps_from_s3}
\title{Read processed BPS data from S3}
\usage{
read_bps_from_s3(
  vintage = "latest",
  variables = NULL,
  years = NULL,
  geoids = NULL,
  s3_bucket = "cori.data.bps",
  s3_path_prefix = ""
)
}
\arguments{
\item{vintage}{Passed to [get_building_permits()].}

\item{variables}{Passed to [get_building_permits()].}

\item{years}{Passed to [get_building_permits()].}

\item{geoids}{Passed to [get_building_permits()].}

\item{s3_bucket}{Ignored. No longer configurable in the public API.}

\item{s3_path_prefix}{Ignored. No longer configurable in the public API.}
}
\value{
A data frame. See [get_building_permits()] for details.
}
\description{
`r lifecycle::badge("deprecated")`

`read_bps_from_s3()` is deprecated. Use [get_building_permits()] instead.
}
\seealso{
[get_building_permits()]
}
