% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bps_codebook.R
\name{get_bps_codebook}
\alias{get_bps_codebook}
\title{Get the cori.data.bps variable codebook}
\usage{
get_bps_codebook()
}
\value{
A data frame with columns: \code{variable}, \code{label},
  \code{unit}, \code{nominal}, \code{notes}.
}
\description{
Returns documentation for all variables in the processed BPS output.
}
\examples{
get_bps_codebook()

}
\seealso{
\code{\link{read_bps_from_s3}}
}
